Changed the type of migration_time from int to float

+++ neat/locals/overload/statistics.py
@@ -34,7 +34,7 @@ def loess_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
@@ -59,7 +59,7 @@ def loess_robust_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
@@ -84,7 +84,7 @@ def mad_threshold_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
@@ -107,7 +107,7 @@ def iqr_threshold_factory(time_step, migration_time, params):
      :type time_step: int,>=0
 
     :param migration_time: The VM migration time in time seconds.
-     :type migration_time: int,>=0
+     :type migration_time: float,>=0
 
     :param params: A dictionary containing the algorithm's parameters.
      :type params: dict(str: *)
